+++ primitives/Cargo.toml
@@ -6,7 +6,17 @@ edition = "2018"
 license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 
 [dependencies]
-codec = { version = "1.3.5", package = "parity-scale-codec", features = ["derive"] }
+codec = { version = "1.3.5", package = "parity-scale-codec", default-features = false, features = ["derive"] }
+sp-application-crypto = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
+sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
 
 [dev-dependencies]
 hex-literal = "0.3"
+
+[features]
+default = ["std"]
+std = [
+	"codec/std",
+	"sp-application-crypto/std",
+	"sp-runtime/std",
+]
